fix: missing alerts in incidents list sidebar when in playback
