+++ src/constants/alerts.js
@@ -1,3 +1,11 @@
+export const ALERT_STATUSES = {
+  NOT_ALERTING: 'NOT_ALERTING',
+  WARNING: 'WARNING',
+  HIGH: 'HIGH',
+  CRITICAL: 'CRITICAL',
+  NOT_CONFIGURED: 'NOT_CONFIGURED',
+};
+
 export const CONDITION_DOMAIN_TYPE = {
   domain: 'AIOPS',
   type: 'CONDITION',
